Created Optimisation module and reorganised WorkFlow module
