+++ ceasiompy/SkinFriction/skinfriction.py
@@ -169,7 +169,7 @@ def add_skin_friction(cpacs_path,cpacs_out_path):
     analyses_xpath = '/cpacs/toolspecific/CEASIOMpy/geometry/analysis'
 
     # Requiered input data from CPACS
-    wetted_area = get_value(tixi,analyses_xpath + '/wettedArea')
+    wetted_area = 1000#get_value(tixi,analyses_xpath + '/wettedArea')#----------------------------------MODIFIED
 
     # Wing area/span, default values will be calated if no value found in the CPACS file
     wing_area_xpath = analyses_xpath + '/wingArea'
@@ -284,8 +284,8 @@ if __name__ == '__main__':
     MODULE_DIR = os.path.dirname(os.path.abspath(__file__))
     cpacs_path = os.path.join(MODULE_DIR,'ToolInput','ToolInput.xml')
     cpacs_out_path = os.path.join(MODULE_DIR,'ToolOutput','ToolOutput.xml')
-
-    check_cpacs_input_requirements(cpacs_path)
+    print("-------------------------------------NO CHECK FUNCTION------------------------------------------")
+    #check_cpacs_input_requirements(cpacs_path)
 
     add_skin_friction(cpacs_path,cpacs_out_path)
